Decouple updating local knownVC from propagation
Before, propagation took care of advancing the local knownVC entry,
so it needed to happen often enough for local clients to see their own
updates.

However, since we're sending a message to the propagation vnode anyway
to append the blue transaction, we can piggyback the new known time with
the message.

Now that propagation is not needed to advance the local clock, we can
replicate less frequently, which will put less load on the system.