Move propagate timer to propagate_vnode
With the move to uniform propagation, the call to PROPAGATE_TX might
be expensive (quadratic w.r.t. number of replicas). We want to keep
back-pressure to a minimum, so the timer to run this function should
only be re-armed after PROPAGATE_TX finishes.

Before, the timer was placed before grb_main_vnode. We need to compute
the new cut-off time for knownVC, and this vnode has local access to the
prepared queue that contains the necessary information to compute the
time.

The node would, on timer fire, compute knownVC[d] and forward it to the
propagation vnode, who would update the local knownVC, as well as use
the cut-off time for propagation.

Since main_vnode uses command/3, the call to the prop. vnode is async,
and as a result is not aware of the time it could take to finish. We
don't want to use sync_command/3 here since it could block main_vnode
for long, causing it to delay blue prepare/decides, which are in the
hot path for clients.

Now, the timer is placed in the propagation vnode, which will call
main_vnode in turn to compute the cut_off time. This will block main
vnode, but only for a very short amount of time, since this computation
is fast. (this computation is done on pull, not on push). Since the timer
is now local to the propagation vnode, it can delay the re-arming of the
timer until the propagation is done. This should lower the amount of
outstanding messages in the vnode queue.