Register uniform barriers as callbacks in the vnode, don't wait
Since the timer to recompute the different uniform barriers has the same
frequence as the interval to re-compute the uniformVC, might as well do
it all in the same place. Clients register their timestamp in an ordered
table, and the vnode lifts the appropriate barriers as the uniformVC is
re-computed.