Advance uniformVC in single-dc mode
In multi-dc mode, uniformVC is recomputed when sending our local
stableVC to a remote replica, or when receiving a broadcast_clock
request. As such, in single-dc mode, our uniformVC will never advance.

Since we use uniformVC to compute the initial snapshotVC of
transactions, a new transaction will always start with an empty clock,
instead of seeing previous updates.