Replace uniform barrier ets w/ normal orddict
Scanning the orddict is much faster than ets:select, since the entries
are ordered in ascending time, we can just traverse the list of entries
up until we find a key that is over our time. Also, we can match on an
empty barrier list, which helps significantly in the common case where
there are no barriers (since we bail out during match, instead of
scanning and iterating over an empty list).

Reduces from 4-10ms (!) to 0-1 microseconds