Recompute local uniformVC on stableVC update
We were waiting until propagation to re-compute our local uniformVC (and
lift uniform barriers). This introduces a delay since we can re-compute
our local entry as soon as we have a new local stableVC. Also,
recomputing the uniformVC on propagate doesn't make much sense, since we
don't use it at that time.

We're still re-computing the uniformVC as soon as we receive a stableVC
from a remote replica.