Fix local broadcast crash (introduced in 10878c1)
When the knownVC was split into several entries, we introduced the
possibility of a race condition where the local stableVC was recomputed
before we received any heartbeats from remote replicas.

grb_propagation_vnode:known_time_internal/2 assumed that we already had
a valid clock time for the requested replica, since it's the
responsibility of the caller to ensure that the remote replica exists.
However, there's a gap of time between when a replica is known and
connected, and when that replica sends a heartbeat message. During that
gap, calling known_time_internal could lead to a badarg crash, since the
key for the remote replica doesn't exist yet.