Use ETS commit logs for remote transactions
Benchmarks show that one of the big differences between Cure and Uniform
Replication comes from having to store remote transactions in
committedBlue, so we can replicate them later in case of replica
failure.

Simply sending a message to grb_propagation_vnode, and not storing the
transaction, is enough to hurt performance. Changing the commitLog
structure for remote transactions to a public ETS table should avoid
this message, so that grb_main_vnode simply copies the data to the ETS
table, instead of going through the other vnode process.

Using ETS tables for the local commitLog structure is too much, as it
causes quite a big slowdown when replicating transactions. With the
current orddict-based structure, we replicate 5-10 transactions every
time the replication timer fires. With a public ETS table
implementation, this number grows to 1K transactions, which causes the
process to hang for a long time on gen_tcp:send/2. Since Cure also uses
an orddict to replicate transactions, the local log is not a sufficient
bottleneck to find an alternative for it.