Introduce a cache for RedTs / CommitVC
During CERTIFICATION_CHECK, we have to check that the red version in the
transaction's writeset is newer than any other red version in the opLog.
In addition, we have to update the prepareVC with the commitVC of any
such entry. Since the opLog only contains a grb_version_log structure,
it becomes hard to check against that.

The new ETS table keeps track of the max red timestamp for a certain
key, so that checking against that is a simple ETS lookup. To aggregate
the commitVCs, we also include a list of commitVCs in the ETS table, so
we can also quickly iterate over them. This list will only be copied
into the process if needed, that is, if we find that the transaction
should abort, we skip the aggregation process.

To limit the performance penalty of copying a large list of vectors when
we want to aggregate, we could keep only one vector, and aggregate on
commit. However, this would affect _all_ transactions, since blue
transactions would have to do this aggregation, too. That's why we keep
the list capped at VSN_LOG_SIZE elements. When a transaction commits, it
checks if the list of commitVCs is larger, and in that case, will
compact it into a single element-list with the max of all vectors. This
way, most transactions will simply append a clock, update the red entry,
and update the list length, which should be fast. When a transaction
hits the limit, it will also hit it for the normal grb_version_log
structure, so it's already paying that price.