Buffer out-of-order DECISION
It is possible for a replica to receive a DECISION message from the
coordinator before it receives an ACCEPT message from its leader (since
the coordinator only needs to receive a quorum of ACCEPT_ACK before
deciding a transaction).

As such, if we receive an out-of-order DECISION message, we buffer it
until we receive the corresponding ACCEPT message from the leader.

Caveat: in the case of leader failure, we should drop buffered messages.
In any case, when we buffer them, we store the ballot, so we never mix
messages about the same transaction from different leaders.