Rework async_op
Do a tentative check of the partition outside grb_partition_replica. If
the partition is ready, proceed with the write. In most cases, the
partition will be ready and we can avoid copying the client payload to
grb_partition_replica, and we can return to the client immediately.

If the partition is not ready, go through the old path.