Fix potential missed red updates
Previously, red transactions could enter a state where its updates would
never be delivered, or a partition would deliver red transactions in a
bad order. When we received a decision message for a transaction that
wasn't yet accepted, we would buffer the decision until we received the
matching accept. However, because we didn't log its commit timestamp,
other transactions or heartbeats with a higher commit timestmap could be
delivered (since paxos_state doesn't know about the previous
transaction).

As a result, the vnode could deliver heartbeats or newer transactions
and advance its LastDelivered timestamp. This would cause the original
transaction to sit in the paxos_state queue forever (or until it was
pruned afterwards), without ever being delivered to the partition.