Move to a single inter-dc connection per partition
Remove shackle, use our own gen_servers wrapping a socket. Senders don't
need to go through the gen_server to send a message, though, since they
can use the connection handler.

Now that we have a connection per partition, we can send the partition
and our replica identifier at the beginning of the conneciton using a
ping message. In future messages, we can avoid both of these fields,
saving on bandwidth.

For forwarded messages (heartbeat / tx), we can include a new field in
the message telling the remote DC to use that field as the replica id of
the sender.