Change red heartbeat rate
Previously, we waited `red_heartbeat_interval` _after_ the previous
heartbeat was done, which doesn't make sense in high latency scenarios.
If the commit latency is of 100ms, and the interval is 1ms, this means
that partitions only receive a heartbeat every 101ms. This means that
a transaction could wait an average of 101/2 ms for a heartbeat to be
delivered.

Now we start a red heartbeat every `red_heartbeat_interval` ms, and keep
a list of currently active heartbeats that the red timer coordinator has
to keep alive. If we have a RTT of 100ms, and 1ms interval, this means
that we would have roughly 100 active heartbeats at any given time.