Include payload on decide message, avoid buffering
If a replica receives a decision before an accept, it would need to wait
for the accept, since it is the message that contains the payload.

Now, we include the readset/writeset in the decision message, so that
replicas that receive a decision for a not-yet accepted transactions can
still proceed.