Make remove_from_prepared/3 more efficient
Add a secondary index from Tx -> PrepTime so we don't have to scan the
entire PepareBlue table when we want to remove a transaction.

Since the commit time might be different from the prepare time, we
couldn't know the entire key of the transaction in the prepared table.
Now we store the prepare time in the index so we can delete the entry
directly, without scanning the table.