Remove from prepared after inserting on commit log
Fixes some bug where a heartbeat could be sent in the time between
removing from preparedBlue and inserting into committedBlue. If this
happened, a partition and its replicas would advance knownVC past the
commit time of the transaction without seeing it. This would result in
the transaction never being replicated by other replicas.