Fix issue in multiple concurrent reads from the same transaction
When a transaction issues concurrent reads to a single partition, if
those two requests fall in the same oplog_reader, they might overwrite
each other on the simple_waiting_reads field, as only the transaction id
is used as identifier. This could also cause issues where the process
tries to handle a `retry_partition_wait` message, but the entry for the
transaction was removed during a previous try.

Also, add comment on the potential for the same issue to appear with
concurrent multi-key requests.