Fix duplicate detection on get_next_ready/2
If we deliver a transaction with the same commit time as another
transaction still sitting on prepareRed, we could update lastDelivered
past the commit time in preparedRed, preventing it from ever being
delivered.

The same applies to reserve_decision, if we receive the decision for a
transaction before the accept, in the case that the accept message would
cause the issue described above.