Route decision messages through the leader
Sending decision messages to the followers from the coordinator might
cause an issue where a transaction gets delivered before another with a
lower commit timestamp gets accepted, which causes some transactions to
never be delivered. By routing decisions through the leader, we ensure
that FIFO order is respected: if a follower receives a decision for
transaction, it must have received an accept for any previously decided
transactions.