Enable blue stall check
If a transaction has sit for too long of preparedBlue, it might mean
that the client has errored, and will never send a decide. If this
happens, it could stall the system (knownVec never makes any progress).

We should check every X ms if this has happened, and if so, remove the
transaction from preparedBlue.