Implement passive red replication
Following discussion of the commit protocol, it's not enough to proxy
decision messages through the leader. Only the leader knows when it's
safe to deliver a transaction, since only a quorum of replicas is needed
to commit a transaction.

With this change, the delivery precondition is only checked at the
leader, and deliver messages are sent to the followers. To optimize
things, and avoid two messages in the commit case (one decision, one
deliver), we only forward decision messages for aborted transactions,
and wait until delivery to let followers know about the decision for a
transaction. This also avoids the decision-before-accept scenario, due
to FIFO guarantees.