Enable strong heartbeats only when partition is idle
Strong heartbeats are expensive, since they are equivalent
to a read-only strong transaction that always commits. If
we're sending a heartbeat every X ms, that's (1000 / X)
transactions per second and per partition being performed.

To avoid paying the price, we look if the process is currently
certifying or delivering transactions. If the system is not
doing anything, we schedule a heartbeat. If, on the other hand,
the process is preparing or delivering transactions, we know clients
are submitting transactions, and thus we know the system will advance
even if we don't send heartbeats.

Since delivery is checked every 1ms, in an idle system the final
result is similar, performing a heartbeat every (1 + X) ms, until the
system starts processing client transactions. Heartbeats stop being
scheduled until the system finds itself not preparing / delivering
transactions for an amount of time.