Delay sending abort messages to followers
The leader can wait for a while before sending abort messages to followers.
In the normal case, followers don't need to learn about aborted transactions,
since they don't execute any delivery preconditions. This allows us to save
an exchaned message during commit if we know the transactions is aborted.

During recovery, it's important that the new leader knows about aborted
transactions, otherwise it won't be able to deliver new transactions.

A solution to this problem is to retry transactions that have been sitting
in prepared for too long.