Fix stat match when key is nested tuple
If the key for a stat contains a nested tuple like `{a, {b, c}}`, the
match spec body will fail, because it will only wrap the outer tuple,
like `{ {a, {b, c}} }`, when ETS would expect `{ {a, { {b, c} } } }`.
This is because match specs expect _all_ tuples to be surrounded by the
double bracket parenthesis. A way to fix it, and avoid the hacky
`is_tuple/1` workaround, is to simply wrap the key with `const`, which
will preserve the key structure.