Remove out-of-order decision handling
With the old version of the protocol, a follower could receive a
decision for a transaction from the coordinator before receiving an
accept from the leader. This could happen since the coordinator only
needed a quorum of accept_ack messages, and would talk directly to
followers when sending decision messages.

Now, decision messages are proxied through the leader, which gives us
FIFO guarantees: if a follower receives a decision message from the
leader, it must have received an accept. Maybe this doesn't hold across
recoveries, but in that case, the ballot will have changed.