Change get_next_ready/2 implementation
Old implementation performed a full-table scan to find the first
committed transaction. We can merge that find with the
prep_committed_between/2 check, just by walking from the first
transaction that is after `LastDelivered`. If we find a prepared commit
before we find a decided commit, we know that no transaction is ready to
be delivered.

With that done, the only thing that remains is to check against other
prepared transactions with the same commit timestamp.