Simplify node-checking logic on coordinator
Checking on the node before calling the coordinator might be error-prone
for the caller, as it has to repeat the arguments twice. We now check
the node at the callsite instead, and use spawn_request directly instead
of going through erpc, since that will perform an extra check against
node().