Optimize riak_core_vnode_master:command/{3,4}
Riak core will use a proxy process even if the target Erlang node is in
the same physical machine. For our use case, where most of the time
we're sure we're in the same erlang node, it's a nuisance. We can
register the vnode pid and use that directly to communicate with the
vnode.

Instead of changing every handle_command for handle_info, we can hook up
into riak_core's implementation and use gen_fsm_compat to deliver our
requests directly to the process without going through a proxy server.
This way we can also avoid calling node/0 for every call, and we avoid
allocating a tuple and a list for every call.