Clean pending strong client operations at decision/delivery time
Before we were cleaning the pending operations at prepare time at the
leader, and at accept time at the followers. Now, if the transaction is
an abort, we remove them when we learn about the decision (at the
leader), or when we receive a learn_abort message (at the followers).

For committed transactions, it will be the oplog_vnode the one who will
clean up the pending operations when it applies the updates to the
storage.