Remove PrepareVC from ACCEPT_ACK, return only prepare timestamp
The leader only updates the commit timestamp to the red entry, while the
rest of the snapshot vector doesn't change. Since the coordinator can
keep the original snapshot vector around, it doesn't make sense to
return the entire vector in the message, the coordinator can perform the
max over the commit timestamps only.