Remove writeset from DELIVER message
We were sending the writeset with the delivery message, but this is not
needed, since followers already have it stored in their state (due to
FIFO, if they receive a DELIVER(tx), they must have seen ACCEPT(tx),
which contained both readsets and writesets.

The followers now retrieve the writeset from local storage right before
delivering it to the oplog vnode.