Remove CommitVC from decision / delivery
For decision, it's enough to return the commit timestamp, since the
leader has the prepare vector that was used to prepare the transaction.

For delivery, since followers also have the prepare vector, they only
need the commit timestamp to update the state. And since the delivery
message has the commit timestamp of all the entries in a field, we can
let the receiver use it as the commit timestamp of all incoming
transactions.

This should reduce the size of messages on the wire, as we're saving
around 140 bytes for the 3-site configuration. Each site entry in the
vector clock consumes around 30-35 bytes depending on the name of the
site.